requirements.txt
mysql-connector-python
discord.py
requests
beautifulsoup4
pandas
numpy
scikit-learn
matplotlib
pytest
python-dotenv
selenium
aiohttp
lxml
loguru
transformers 
torch 
torchvision 
torchaudio
webdriver-manager
textblob
vaderSentiment
fastapi
uvicorn
plotly
joblib
alpaca-trade-api
yfinance
pyyaml
dataclasses
PyQt6